(#132) Moved client namespace to Azure.Data.Mobile
