+++ dotnet/client/src/Azure.Data.Mobile/Azure.Data.Mobile.csproj
@@ -3,10 +3,12 @@
   <PropertyGroup>
     <TargetFramework>netstandard2.0</TargetFramework>
     <LangVersion>8.0</LangVersion>
+    <Version>5.0.0</Version>
   </PropertyGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\client\src\Azure.Data.Mobile\Azure.Data.Mobile.csproj" />
+    <PackageReference Include="Azure.Core" Version="1.3.0" />
+    <PackageReference Include="System.Text.Json" Version="4.7.2" />
   </ItemGroup>
 
 </Project>
