+++ dotnet/client/src/Azure.Data.Mobile/MobileDataClientOptions.cs
@@ -1,0 +1,15 @@
+﻿using Azure.Core;
+using System.Text.Json;
+
+namespace Azure.Data.Mobile
+{
+    public class MobileDataClientOptions : ClientOptions
+    {
+        /// <summary>
+        /// The SerializerOptions used by the serialization library.  The default is
+        /// normally good, unless you adjust on the service too.
+        /// </summary>
+        public JsonSerializerOptions JsonSerializerOptions { get; set; }
+            = new JsonSerializerOptions { PropertyNamingPolicy = JsonNamingPolicy.CamelCase };
+    }
+}
