+++ dotnet/client/src/Azure.Data.Mobile/Properties/InternalsVisibleTo.cs
@@ -1,0 +1,3 @@
+﻿using System.Runtime.CompilerServices;
+
+[assembly: InternalsVisibleTo("Azure.Data.Mobile.Test")]
