--- dotnet/client/src/Azure.Mobile.Client/Properties/InternalsVisibleTo.cs
@@ -1,3 +1,0 @@
-﻿using System.Runtime.CompilerServices;
-
-[assembly: InternalsVisibleTo("Azure.Mobile.Client.Test")]
