+++ dotnet/client/test/Azure.Data.Mobile.Test/Helpers/HttpAssert.cs
@@ -1,0 +1,17 @@
+﻿using Azure.Core;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
+
+namespace Azure.Data.Mobile.Test.Helpers
+{
+    internal static class HttpAssert
+    {
+        internal static void HeaderIsEqual(string headerName, string headerValue, Request request)
+        {
+            string actual;
+            var isPresent = request.Headers.TryGetValue(headerName, out actual);
+            Assert.IsTrue(isPresent);
+            Assert.IsNotNull(actual);
+            Assert.AreEqual(headerValue, actual);
+        }
+    }
+}
